+++ packages/git-service-koa/lib/index.d.ts
@@ -1,8 +1,9 @@
 /// <reference types="koa" />
 import { LogicController } from "git-service";
 import { Middleware } from "koa";
-export declare function createMiddleware(controller: LogicController, options?: IMiddlewareOptions): Middleware;
-export interface IMiddlewareOptions {
+export * from "git-service";
+export declare function createKoaMiddleware(controller: LogicController, options?: IKoaMiddlewareOptions): Middleware;
+export interface IKoaMiddlewareOptions {
     /**
      * Where to store proxy in `Context.state`. Defaults to `"service"`.
      */