+++ packages/git-service-koa/lib/index.js
@@ -1,11 +1,17 @@
 "use strict";
+function __export(m) {
+    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
 Object.defineProperty(exports, "__esModule", { value: true });
 const git_service_1 = require("git-service");
-function createMiddleware(controller, options = {}) {
+__export(require("git-service"));
+function createKoaMiddleware(controller, options = {}) {
     const keyName = options.keyName || "service";
-    const autoDeploy = options.autoDeploy || true;
+    const autoDeploy = "autoDeploy" in options ? options.autoDeploy : true;
     return async (context, next) => {
         const service = context.state[keyName] = git_service_1.createService(controller, context.url, context.method.toUpperCase(), context.headers, context.req);
+        // Link service state to context state.
+        service.onRequest.addOnce((request) => request.state = context.state);
         await next();
         if (autoDeploy) {
             try {
@@ -24,5 +30,5 @@ function createMiddleware(controller, options = {}) {
         }
     };
 }
-exports.createMiddleware = createMiddleware;
+exports.createKoaMiddleware = createKoaMiddleware;
 //# sourceMappingURL=index.js.map